fixing windows build (#211)
* fixed warnings

* enabled openssl

* fixed PR issues, added zlib to compile

* fixed PR issues

* fixed PR issues

* fixed PR issues

* merge with main repo

* fixing PR issues

* fixing PR issues

* replaced ff with tbb

* removed ff and config option for mpmc

* replaced mimalloc with tbb_malloc

* alsmost fixed windows build w/o ssl

* removed .vcprojs, .slns

* fixed linux build

* fixed msvc warnings

* fixed liniking of tbb_malloc_proxy

* partial fix for session_test crash

---------

+++ include/fix8/field.hpp
@@ -96,7 +96,7 @@ struct RealmBase
 		if (_dtype == dt_set)
 		{
 			const T *rng(static_cast<const T*>(_range)), *res(std::lower_bound(rng, rng + _sz, what));
-			return res != rng + _sz ? res - rng : -1;
+			return res != rng + _sz ? static_cast<int>(res - rng) : -1;
 		}
 		return 0;
 	}
